+++ haskell/ngx-export-distribution/NgxExport/Distribution.hs
@@ -39,9 +39,7 @@ import Distribution.Types.Library
 import Distribution.Verbosity
 import Distribution.Pretty
 import System.FilePath
-import Control.Exception
 import Control.Arrow
-import Control.Monad
 import Data.Maybe
 
 -- $usage
@@ -101,9 +99,8 @@ import Data.Maybe
 --
 -- Building is a bit cumbersome: it expects explicit option /--prefix/ at the
 -- configuration step (which will be interpreted as the first part of the
--- /rpath/ by utility /hslibdeps/), and requires explicit ghc option /-o/ at
--- the build step which is as well used by /hslibdeps/ as the name of the
--- target library.
+-- /rpath/ by utility /hslibdeps/) and explicit ghc option /-o/ at the build
+-- step which is as well used by /hslibdeps/ as the name of the target library.
 --
 -- Let's build the example with commands /cabal v1-configure/ and
 -- /cabal v1-build/ (the /v2-/commands should probably work as well).
@@ -149,13 +146,6 @@ import Data.Maybe
 -- - behavior of Cabal commands other than /configure/, /build/ and /clean/ is
 --   not well defined.
 
-data LibNameNotSpecified = LibNameNotSpecified
-
-instance Exception LibNameNotSpecified
-instance Show LibNameNotSpecified where
-    show = const "Error: the library name was not specified, \
-                 \the name must be passed in ghc with option -o"
-
 hslibdeps :: Program
 hslibdeps = simpleProgram "hslibdeps"
 
@@ -168,10 +158,8 @@ patchelf = simpleProgram "patchelf"
 --
 -- - /-dynamic/, /-shared/, /-fPIC/, /-flink-rts/,
 -- - all arguments listed in /ghc-options/ in the Cabal file,
--- - all arguments passed in option /--ghc-options/ from command-line.
---
--- Requires that the arguments contain /-o path/ for the path to the shared
--- library to build.
+-- - all arguments passed in option /--ghc-options/ from command-line,
+-- - if arguments do not contain /-o path/ so far, then /$pkg.hs/, /-o $pkg.so/.
 --
 -- Returns the path to the built shared library.
 buildSharedLib :: Verbosity                         -- ^ Verbosity level
@@ -200,13 +188,17 @@ buildSharedLib verbosity desc lbi flags = do
                   ) (Nothing, False) $
                       buildProgramArgs flags ++
                           map (second pure) configGhcOptions
-    when (isNothing lib) $ throwIO LibNameNotSpecified
+        (lib', extraGhcOptions) =
+            maybe (let name = unPackageName $ pkgName $ package desc
+                       nameSo = addExtension name "so"
+                   in (nameSo, [addExtension name "hs", "-o", nameSo])
+                  ) (, []) lib
     ghcP <- fst <$> requireProgram verbosity ghcProgram (withPrograms lbi)
     let ghcR = programInvocation ghcP $
             ["-dynamic", "-shared", "-fPIC", "-flink-rts"] ++
-                map snd configGhcOptions
+                map snd configGhcOptions ++ extraGhcOptions
     runProgramInvocation verbosity ghcR
-    return $ fromJust lib
+    return lib'
 
 -- | Patches the shared library and collects dependent Haskell libraries.
 --