+++ haskell/ngx-export-distribution/Changelog.md
@@ -1,6 +1,8 @@
 ### 0.3.0.0
 
 - GHC option *-flink-rts* was added to the list of default GHC options.
+- Add GHC options *&dollar;pkg.hs -o &dollar;pkg.so* when building without
+  option *-o* specified.
 - More robust naming of paths to build artifacts.
 
 ### 0.2.0.0